refactor : MixpanelEventListener 사용
