+++ infrastructure/src/main/java/org/depromeet/spot/infrastructure/mixpanel/repository/MixpanelRepositoryImpl.java
@@ -2,9 +2,9 @@ package org.depromeet.spot.infrastructure.mixpanel.repository;
 
 import java.io.IOException;
 
-import org.depromeet.spot.domain.mixpanel.MixpanelEvent;
 import org.depromeet.spot.infrastructure.mixpanel.property.MixpanelProperties;
 import org.depromeet.spot.usecase.port.out.mixpanel.MixpanelRepository;
+import org.depromeet.spot.usecase.service.event.MixpanelEvent;
 import org.json.JSONObject;
 import org.springframework.stereotype.Component;
 
@@ -32,14 +32,18 @@ public class MixpanelRepositoryImpl implements MixpanelRepository {
     // mixpanelEvent는 eventName(이 단위로 이벤트가 묶임)
     // distinctId는 사용자를 구분하는 데 사용됨.
     @Override
-    public void eventTrack(MixpanelEvent mixpanelEvent, String distinctId) {
+    public void eventTrack(MixpanelEvent mixpanelEvent) {
         try {
 
             // 믹스패널 이벤트 메시지 생성
             MessageBuilder messageBuilder = new MessageBuilder(mixpanelProperties.token());
 
             // 이벤트 생성
-            JSONObject sentEvent = messageBuilder.event(distinctId, mixpanelEvent.getValue(), null);
+            JSONObject sentEvent =
+                    messageBuilder.event(
+                            mixpanelEvent.getDistinctId(),
+                            mixpanelEvent.getMixpanelEventName().getValue(),
+                            null);
 
             // 만든 여러 이벤트를 delivery
             ClientDelivery delivery = new ClientDelivery();
