+++ usecase/src/main/java/org/depromeet/spot/usecase/service/review/CreateReviewService.java
@@ -4,17 +4,18 @@ import java.util.List;
 import java.util.Map;
 
 import org.depromeet.spot.domain.member.Member;
-import org.depromeet.spot.domain.mixpanel.MixpanelEvent;
 import org.depromeet.spot.domain.review.Review;
 import org.depromeet.spot.domain.review.keyword.Keyword;
 import org.depromeet.spot.usecase.port.in.review.CreateReviewUsecase;
 import org.depromeet.spot.usecase.port.out.member.MemberRepository;
-import org.depromeet.spot.usecase.port.out.mixpanel.MixpanelRepository;
 import org.depromeet.spot.usecase.port.out.review.ReviewRepository;
+import org.depromeet.spot.usecase.service.event.MixpanelEvent;
+import org.depromeet.spot.usecase.service.event.MixpanelEvent.MixpanelEventName;
 import org.depromeet.spot.usecase.service.member.processor.MemberLevelProcessor;
 import org.depromeet.spot.usecase.service.review.processor.ReviewCreationProcessor;
 import org.depromeet.spot.usecase.service.review.processor.ReviewImageProcessor;
 import org.depromeet.spot.usecase.service.review.processor.ReviewKeywordProcessor;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -26,13 +27,14 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class CreateReviewService implements CreateReviewUsecase {
 
+    private final ApplicationEventPublisher applicationEventPublisher;
+
     private final MemberRepository memberRepository;
     private final ReviewRepository reviewRepository;
     private final ReviewCreationProcessor reviewCreationProcessor;
     private final ReviewImageProcessor reviewImageProcessor;
     private final ReviewKeywordProcessor reviewKeywordProcessor;
     private final MemberLevelProcessor memberLevelProcessor;
-    private final MixpanelRepository mixpanelRepository;
 
     @Override
     @Transactional
@@ -50,7 +52,8 @@ public class CreateReviewService implements CreateReviewUsecase {
         Member levelUpdateMember = memberLevelProcessor.calculateAndUpdateMemberLevel(member);
 
         // 믹스패널 이벤트(후기 등록 완료) 호출
-        mixpanelRepository.eventTrack(MixpanelEvent.REVIEW_REGISTER, String.valueOf(memberId));
+        applicationEventPublisher.publishEvent(
+                new MixpanelEvent(MixpanelEventName.REVIEW_REGISTER, String.valueOf(memberId)));
 
         return new CreateReviewResult(savedReview, levelUpdateMember, review.getSeat());
     }
